+++ examples/aggregation/_01_mobilised_aggregator.py
@@ -73,9 +73,14 @@ agg.aggregate(data, wb_dmos_mask=data_mask)
 # %%
 # The resulting :class:`pandas.DataFrame` containing the aggregated data contains one row for every group.
 # In this case, there is only one participant and day, so the resulting dataframe contains only one row.
-agg.aggregated_data_
+agg_data = agg.aggregated_data_
+agg_data
 
 # %%
+# .. warn:: To exactly match the expected output of the original Mobilise-D R-Script, the two stride length parameters
+#           would need to be converted to cm and all values rounded to 3 decimals.
+#           This is not done in the Python implementation to be consistent with the units across the entire package.
+#
 # Comparison with R aggregation script
 # ------------------------------------
 # The outputs of this aggregation algorithm are analogous to the outputs of the original Mobilise-D R-Script, using